intel-pfr: convert bitmap_size to bit unit in decompression_erase()
The loop that processes the erase bitmap is using bit-level indexing
(bit_in_bitmap), but the variable `bitmap_size` was previously in
bytes. To ensure proper bounds checking, the `bitmap_size` is now
converted to bits.

This prevents potential early loop exit and ensures that
all required bits in the erase bitmap are properly processed.

Signed-off-by: Kane-Chen-AS <kane_chen@aspeedtech.com>
Change-Id: I4e0f52d50513e9a4e2075ac82977bdff6f4a4fd3